pyproject.toml
[project]
name = "archivos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "atlassian-python-api>=4.0.4",
    "dotenv>=0.9.9",
    "markdownify>=1.1.0",
]
